Refactor smithy middleware and steps
This refactors smithy "middleware" and "steps" to be more generic.

First, the previous naming for SmithyMiddleware was confusing and the
typing definitions weren't ideal. SmithyMiddlware has been removed in
favor of defining these things as aliases of the more generic middleware
types. This makes defining middleware less cumbersome in general as the
type of the middleware will be derived from the handler it's wrapping
without being explicitly declared. This also makes it possible to
register a middleware definition on any step it's compatible with at a
typing level as they're no longer arbitrarily tied to a step.

Next, the concept of a "step" as an object has been removed in favor of
a more generic entry/collection system that could be reused in other
contexts. A step is now type alias for a collection of specific
middleware entries of a particular input type that correspond to that
particular step.