Test middleware and http
Adds basic testing for HTTP interfaces, the CRT HTTP implementation, and
the Smithy middleware implementation. This also fixes any bugs found
during testing.

This updates the project to use mypy configuration in the setup.cfg
instead of relying on invocations of mypy to set the appropriate flags.
Mypy should always be ran in strict mode and bypassing checks should be
narrowly scoped to the setup.cfg or applied to specific lines if a
reasonable scope can't be defined.

This updates our coverage configuration to re-include ignore lines with
no cover and also adds an additional flag to ignore crt-callback. It
appears that the interfaces the CRT is using to callback into the code
we've written does not get marked as being ran by coverage despite
actually being executed.