Generate complete python package
This updates the generated code to be a complete python package.
In attempt to follow most recent standards, a pyproject.toml is
generated to allow for potential usage with other build tools, though
it currently isn't configurable. Since setuptools doesn't yet read
this, a complete setup.cfg is also generated.