Fix as and from dict for collections/maps/unions
This fixes an issue in the `as_dict` and `from_dict` methods where
lists, sets, and maps were never being handled correctly, and
unions weren't being handled correctly in `from_dict`.

Lists, sets, and maps now all generate private `from_dict` and
`as_dict` methods that are used to ensure members are converted.
Similarly, a `from_dict` method is generated for unions.

For lists, sets, and maps the helper methods are not generated if
no conversion is necessary.