Move pre-shape-generation steps to dedicated hook
We have a few bits of code generation that need to happen before we
do the shape code generation. Previously we were leveraging the
service generation hook, but then it got moved to being called after
the rest of the shapes were generated. We had to move things into
temporary locations to make things work, but now there's an explicit
hook before shape generation we can use.