Fix SymbolVisitor handling of union members
Union members get actual classes generated for them, but we were
treating them like we treat members of other shapes in the
SymbolVisitor - namely just forwarding the symbol of the target.
This isn't what we actually want, so this fixes that to return
a symbol representing the generated union member type. The union
generator was also updated to account for this.