Write protocol test skeletons (input, command, assertions)
As part of these changes, some test skeletons are written. A basic
implementation of a `NodeVisitor` is added, which is used to convert
from node values to python values. A small change was made to
PythonWriter to skip the import of operation symbols, since they are
methods of the service object.