Write protocol test skeletons (input, command, assertions) (#92)
As part of these changes, some test skeletons are written. A basic
implementation of a `NodeVisitor` is added, which is used to convert
from node values to python values. A small change was made to
PythonWriter to skip the import of operation symbols, since they are
methods of the service object. A few changes were made to shape
codegen, such that the sparse trait is handled correctly. Import
relativization was removed from code in the `tests` module.