Put default http config directly on the config object
This removes the indirection of the endpoint params. This lets us be
compliant with the more complex endpoint resolution rules without
actually implementing it yet (or requiring customers to buy in). At
operation generation time we can just manually wire this up. If the
more advanced traits are present, we can upgrade.