Refactor config generation
This refactors config generation so that we can produce a better type
signature and so that setting and changing default values is easier.

Previously default values had to be set by adding some default plugin
to set it after creation, but this had two major problems. The first
is that it means that the typing for non-nullable properties had to
be nullable anyway so that they could be set later. This is solved
now because we're setting them directly in the initializer.

The second problem was that it meant we would always import those
plugins and all their transative dependencies regardless of whether
they would be used or not. This is solved by using code sections,
which isolate the addition of dependencies so that if the section
is overwritten the depdencies can also be removed.

Another problem this solves is the problem of the default endpoint
resolver. The way it was being created was super ugly. Rather than
having some file you're expected to write an endpoint resolver to
that will then be imported, now you can just modify the initializer
section to use the resolver you want.