Split up smithy-python
This refactors smithy-python into two packages: smithy-core and
smithy-http, separating concerns to bring things more in line with
the split other sdks are moving towards. This allows non-http
clients to exclude http features altogether. It also makes awscrt
and aiohttp into optional dependencies.

Additionally, this attempts to break out async components into their
own sections so that it will be more feasible to provide a sync
client later on if necessary.

Aside from ading in checks for the optional dependencies, the actual
logic has not changed, only the locations of bits of code.