Update base python to 3.12
Upgrading to python 3.12 provides a ton of benefits both in authoring
and in runtime performance. Of particular note is the new type param
syntax, which will massively improve the readability of the
relatively large amount of generic typing that we make use of. In
addition, there's huge performance gains in asyncio as a whole.

This does not update any python code, only the configurations.