Switch to pyright for type checking
Mypy is agonizingly slow to update. PEP695, for example, has been
approved for nearly a year and yet it isn't supported. This is really
a problem for us because we want to be able to use the latest
features. Pyright not being bound to the built-in python parser lets
it be unburdended by some of the version support issues that mypy
suffers from.

The one downside is that it uses typescript to work. Not ideal. But
it's not like our customers will have to use typescript so it really
doesn't matter.

In the fullness of time we may switch back, after mypy supports 3.12.
Of course if 3.13 has great new typing features we want and we update
our base version to that, then pyright is again likely to update
much faster.