Add structured document type
This updates the document type to split it into two parts. The first
part is a generic JSON-like type which is expected to be fully
protocol agnostic.

The second part is a wrapping type which will handle serialization
and deserialization, including the possibility of subclasses handling
protocol-specific handling of documents that have attached schemas.
This will also be what is used for botocore compatibility.

The actual serialization and deserialization components aren't yet
implemented, since the interfaces can only be defined once documents
are present.