Introduce streaming operations to HTTP clients
This adds a `stream` operation to the HTTP client interface which
opens a bidirectional-capable stream. This operation is implmented
for the CRT, but not for aiohttp because it doesn't have bidirectional
support.

The reason that this requires an entirely new operation is that a
caller must be capable of immediately sending request frames without
waiting for *any kind of response* and be capable of sending them
at the same time as they are receiving server events. The `send`
method kinda forces them to either wait for the initial responses
or force the caller to guess when it's okay to start sending request
frames. Buffering can be done of course, but that starts to pile on
memory.

Having the implementation return an input stream also has the benefit
of letting the underlying client create their own writer, which the
CRT is tentatively planning to do in leu of wrapping a BytesIO.

Another potentially questionable decision here is the banning of
setting the HTTP request body for streaming requests. An alternative
would be to use that as initial data. But if that initial data is
itself an async reader, you're potentially having to buffer a lot
of data while you wait for the initial data to empty out.