Generate event stream operation method signatures
This updates operation generation to generate event stream operations
with the EventStream type as its return value.

It also updates union generation so that unions contain their own
deserialize functions. This is needed to make the them pass the type
check, but also it is best to have them own as much of that as
possible so that the deserializer function can be left to only
dispatch duty.