Add an asynchronous bytes provider
This adds a class that allows bytes to be asynchronously exchanged.
This primarily serves the purpose of enabling event streaming. Event
streams will create a provider under the hood that will be written
to. That provider will then be passed as the trasnport request body,
which will read it as an async bytes iterable.