Generate RestJson event stream implementation
This updates generic event stream generation with recently introduced
changes and also introduces the concrete implementation for RestJson.

Testing for all of this will be done via protocol tests, and in the
early days manual testing.

Since a lot of this is effectively throwaway code, I was more liberal
with type ignoring and using Any types than I otherwise would be. The
request pipeline is going to be moving to pure python soon^tm, and
the typing issues will be resolved at that time.