Lock Smithy to a static version
This updates our Smithy versioning to a single static version that is
not synced via gradle properties. The version currently locked in is
1.52.0, which is not the latest, but 1.52.1 has a bug that prevents
us from generating protocol tests.

Instead of relying on the version range to keep us up to date, we
will now be relying on dependabot. It can also track all the
different dependency file locations, which removes the need for the
syncing via gradle properties.