Handle empty and incomplete event bytes
This adds in explicit handling for both emtpy and incomplete event
bytes. If nothing is able to be read from the source, event decoders
will return None. If there are bytes there, but they're truncated,
then an explicit error is thrown that wraps what would otherwise
be a `struct.error`. This is only applied for truncations that would
not already be caught by checksum validation.