Safely copy crt clients
The request pipeline currently deepcopies the config since it can be
mutated by plugins. This was causing issues when copying a CRT
client, which has un-copiable components to it. This implements the
deepcopy method to behave better.

It may be necessary in the future to either remove the deepcopy,
or modify this implementation to violate the intent of a deepcopy
somewhat. I'm a bit concerned about not sharing the connection pool.