Buffer async inputs to the crt
The "clever" workaround to the crt taking a sync stream doesn't work.
This effectively reverts the crt binding to the previous state of
buffering everything immediately. This unfortunately means that the
input stream must be closed before anything can be read. This is not
a tenable state for the future.

The ideal resolution is for the crt to support reading from an async
stream. We might have to contribute that support. There may be other
options, like using futures... or sleep... which might be necessary.
But please no.