Close crt bridge stream on end_stream if empty
This is a followup to #432. I think the issue was that end_stream
was getting called when the buffer was already empty, which would
put it in an incomplete state. Unfortunately that fix also left it
in an incomplete (but now working) state. This make sure it doesn't
get into that incomplete state and adds a failing test for that
case.