Document union property key workaround
This adds to the doc strings for property keys and typed properties
a warning about using union types and a description of the
workaround. It's essentially the same thing in four places, but it's
a big enough deal to warrant being pasted all over.

I also added some test cases to make sure it works.