Centralize http binding matching and omit empty payloads
This changes the way members are classified to use a separate binding
matcher. This lets us use the match statement rather than if-else
chains and ensures we're minimizing the number of times we have to
iterate over the member list.

This also adds the ability to select whether a payload should be
omitted if nothing is bound to it.