Use natural dict ordering for member index
This updates schemas to use the natural ordering of a dict to
set member index. Recursive shapes are accommodated by using a None
value to reserve the member's spot in the dictionary until the
value can be actually set.

This has two primary effects. Firstly, it allows the member index
to be confidently used at runtime to actually retrieve a member.
This isn't needed in most cases, but can come in handy occasionally.

This also cuts out a huge amount of boilerplate in the generated
code.